Start gui as early on as possible so that we can display splash screen
We do this by adding the gui and its dependencies to the auto-deploy
directory. We also set felix.fileinstall.noInitialDelay to false in
order to five the auto-deployed bundles the chance to be activate
before the other bundles are installed. This introduces an additional
delay before the actual GUI window is started, but because there is a
splash screen now, having to wait a bit longer is fine.