Load all bundles using the "auto start" method
rather than using the "file install" method for some (notably the
persistence and volatile related bundles and webservice).

"File install" is now only used for bundles that are placed in the
"modules" directory, and by default there are none.

This also allows us to clean up and simplify the build.

It also seems to improve the reliability and makes the launch faster.