Include dependencies section in POM with artifacts needed to run Pipeline as library
It includes everything except the web server, persistence, GUI,
updater and OSGi.

Also deploy a JAR that contains the default logback.xml and optionally
the SimpleAPI class (and has the above mentioned dependencies).

Now a project that wants to use Pipeline as a library can simply
depend on org.daisy.pipeline:assembly and will this way automatically
import all the necessary transitive dependencies.